Fix exact linked agent set test
This test statement tested that the list of agents in the generated
document and the set provided for the test are the same. This means that
the test breaks every time a new detector is added into file-scraper.

It's better to check that every linked agent exists, rather than that
the two generated sets are the same; the next set assertion in the test
already did this.